Fix 32-bit st_ino truncation, getcwd getdents cursor, and /proc/<pid>/cmdline

+++ library/include/stat.h
@@ -7,7 +7,7 @@
 
 struct stat {
         unsigned short st_dev;
-        unsigned short st_ino;
+        unsigned long  st_ino;
         unsigned short st_mode;
         unsigned short st_nlink;
         unsigned short st_uid;
